Step1 구현 + 지난 리뷰 피드백 반영 (#588)
* ArticleId 타입을 int에서 long으로 변경

* Password 타입 추가

* ViewUserDto 타입 추가

* Rename FindArticleQuery to FindArticleUseCase

* SignInController 구현

* 패스워드 타입 및 Exception 추가

* UserId 타입을 long으로 변경

* UserId 타입을 long으로 변경

* 1단계 구현 완료

+++ src/main/java/com/kakao/cafe/article/domain/ArticleId.java
@@ -2,13 +2,13 @@ package com.kakao.cafe.article.domain;
 
 public class ArticleId {
 
-    private final int value;
+    private final long value;
 
-    public ArticleId(int value) {
+    public ArticleId(long value) {
         this.value = value;
     }
 
-    public int getValue() {
+    public long getValue() {
         return value;
     }
 }
